streamline typings
