+++ scripts/generate-docs.js
@@ -8,28 +8,11 @@ const formatter = require('./utils/formatter')
 const compiler = require('./utils/compiler')
 const ejs = require('../packages/wdio-cli/node_modules/ejs')
 const { getSubPackages } = require('./utils/helpers')
+const { PROTOCOLS, PROTOCOL_NAMES, MOBILE_PROTOCOLS, VENDOR_PROTOCOLS } = require('./constants')
 
 const config = require('../website/siteConfig')
 const sidebars = require('../website/_sidebars.json')
 
-const PROTOCOLS = {
-    webdriver: require('../packages/webdriver/protocol/webdriver.json'),
-    appium: require('../packages/webdriver/protocol/appium.json'),
-    jsonwp: require('../packages/webdriver/protocol/jsonwp.json'),
-    mjsonwp: require('../packages/webdriver/protocol/mjsonwp.json'),
-    chromium: require('../packages/webdriver/protocol/chromium.json'),
-    saucelabs: require('../packages/webdriver/protocol/saucelabs.json')
-}
-const PROTOCOL_NAMES = {
-    appium: 'Appium',
-    jsonwp: 'JSON Wire Protocol',
-    mjsonwp: 'Mobile JSON Wire Protocol',
-    webdriver: 'Webdriver Protocol',
-    chromium: 'Chromium',
-    saucelabs: 'Sauce Labs'
-}
-const MOBILE_PROTOCOLS = ['appium', 'mjsonwp']
-const VENDOR_PROTOCOLS = ['chromium']
 const TEMPLATE_PATH = path.join(__dirname, 'templates', 'api.tpl.ejs')
 const MARKDOX_OPTIONS = {
     formatter: formatter,
